codex: address shell_command review feedback (#21142)
Move shell_command remote routing tests into the shell command suite so remote_env remains focused on remote infrastructure coverage.
